==> 5_harmfulemote_to_embedding_mapper.ipynb <==
This file contains code that reads in the harmful emotes we extracted from HasanAbi's chat logs and maps them to the corresponding vector in emote embedding space (we do not need entire embedding space since we are only interested in emotes)

# Reading tsv files into dataframes

import pandas as pd
import numpy as np
import csv  

#read emote embedding space into dataframe. Each row corresponds to an emote where every emote is represented as vector of 500 numbers (made using Word2Vec model)
embeddings_df = pd.read_csv("emote_embeddings.tsv", sep = "\t", encoding='utf-8-sig', header = None, quoting= csv.QUOTE_NONE)
print(embeddings_df)

#read emote labels into dataframe.
labels_df = pd.read_csv("emote_labels.tsv", sep = '\t', encoding='utf-8-sig', quoting= csv.QUOTE_NONE)
print(labels_df)

print(labels_df.shape) 
print(embeddings_df.shape) 


print(labels_df["source"].value_counts())

print(labels_df.dtypes)
print(embeddings_df.dtypes)

#Read harmful emotes extracted from HasanAbi's channel into a dataframe (this includes both HasanAbi and global emotes extracted)
harmful_emotes_df = pd.read_csv("Harmfulemotes_analysis17.csv")
print(harmful_emotes_df)

# Mapping harmful emotes to an embedding vector

emotes_used = harmful_emotes_df["Emote"]
print(emotes_used)
print(type(emotes_used))

emote_and_embedding_df = harmful_emotes_df["Emote"].to_frame()
emote_and_embedding_df["Global emote"] = harmful_emotes_df["Global emote"]
emote_and_embedding_df["Channel emote"] = harmful_emotes_df["Channel emote"]
emote_and_embedding_df["embedding"] = np.nan


index = 0
for current_emote in emotes_used: #go through each emote
    
    # Search for the index of where we find this emote in the labels dataframe
    matching_index = labels_df.index[labels_df['word'] == current_emote].tolist()
    
    if len(matching_index) == 0: #maybe emote is stored in lowercase form
        matching_index = labels_df.index[labels_df['word'].str.lower() == current_emote.lower()].tolist()

    if len(matching_index) == 0: #if emote was not found in labels dataframe
        emote_and_embedding_df.at[index, "embedding"] = "Does not exist"
        index += 1
        continue
    
    #Retrieve the 500 element vector corresponding to this emote using the matching_index
    embedding_vector = embeddings_df.iloc[matching_index[0]].values

    emote_and_embedding_df.iloc[index, emote_and_embedding_df.columns.get_loc("embedding")] = np.array_str(embedding_vector)
    index += 1

print(emote_and_embedding_df)

# Write emote to embedding vector mapping to new csv file

emote_and_embedding_df.to_csv("harmfulemote_with_embeddingvector.csv", index = False)
